# calendar_back/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Calendar API",
        "version": "1.0.0"
    },
    "paths": {
        "/register": {
            "post": {
                "description": "Register a new user",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "username": {"type": "string"},
                                "password": {"type": "string"}
                            },
                            "required": ["username", "password"]
                        }
                    }
                ],
                "responses": {
                    "201": {"description": "User registered successfully"}
                }
            }
        },
        "/login": {
            "post": {
                "description": "Authenticate user and get access token",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "username": {"type": "string"},
                                "password": {"type": "string"}
                            },
                            "required": ["username", "password"]
                        }
                    }
                ],
                "responses": {
                    "200": {"description": "Access token generated successfully"},
                    "401": {"description": "Invalid username or password"}
                }
            }
        },
        "/token": {
            "get": {
                "description": "Refresh access token",
                "responses": {
                    "200": {"description": "Access token refreshed successfully"}
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/users": {
            "get": {
                "description": "Get all users",
                "responses": {
                    "200": {"description": "List of all users"}
                }
            }
        },
        "/events": {
            "post": {
                "description": "Save events for a user",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "events": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "title": {"type": "string"},
                                            "start": {"type": "string", "format": "date-time"},
                                            "end": {"type": "string", "format": "date-time"}
                                        },
                                        "required": ["title", "start", "end"]
                                    }
                                }
                            },
                            "required": ["events"]
                        }
                    }
                ],
                "responses": {
                    "201": {"description": "Events saved successfully"}
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "get": {
                "description": "Get events for a user",
                "responses": {
                    "200": {
                        "description": "List of user's events",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id": {"type": "integer"},
                                    "title": {"type": "string"},
                                    "start": {"type": "string", "format": "date-time"},
                                    "end": {"type": "string", "format": "date-time"}
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/events/delete": {
            "delete": {
                "description": "Delete selected events for a user",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "event_ids": {
                                    "type": "array",
                                    "items": {
                                        "type": "integer"
                                    }
                                }
                            },
                            "required": ["event_ids"]
                        }
                    }
                ],
                "responses": {
                    "200": {"description": "Events deleted successfully"}
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}